Implementaciones 3 y 4 corregidas

+++ Guion4/src/matriz_bit3.cpp
@@ -17,7 +17,7 @@ int Filas (const MatrizBit& m){
 	return m.dimension >> 16;
 }
 int Columnas(const MatrizBit& m){
-	int columnas = m.dimension & (pow(2.0,16.0) - 1);
+	int columnas = m.dimension & (unsigned int)(pow(2.0,16.0) - 1);
 	return columnas;
 }
 bool Get(const MatrizBit& m, int f, int c){
